requirements.txt
asgiref==3.11.0
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-allauth==65.13.1
django-appconf==1.2.0
django-imagekit==6.0.0
django-tinymce==5.0.0
idna==3.11
mpmath==1.3.0
numpy==2.4.1
pilkit==3.0
pillow==12.1.0
python-dateutil==2.9.0.post0
requests==2.32.5
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
urllib3==2.6.3
webencodings==0.5.1
